0.8.0.0 RC6 created resources folder containing all not-python stuff
fixed Image.filter_volumes

refactored utility.list -> myutility.list for a conflict with pandas package
