+++ myutility/exceptions.py
