+++ myutility/remove_name_from_images.sh
