+++ resources/examples/data/bayes_data_import.py
@@ -5,7 +5,7 @@ from Global import Global
 from data.BayesDB import BayesDB
 from data.importer.BayesImporter import BayesImporter
 from data.utilities import *
-from utility.exceptions import SubjectListException
+from myutility.exceptions import SubjectListException
 
 if __name__ == "__main__":
