+++ resources/examples/ot/subjetcs_t1_see_images_1by1.py
@@ -1,6 +1,6 @@
 from Global import Global
 from Project import Project
-from utility.myfsl.utils.run import rrun
+from myutility.myfsl.utils.run import rrun
 
 if __name__ == "__main__":
