+++ subject/SubjectEpi.py
@@ -10,14 +10,14 @@ from group.spm_utilities import SubjCondition
 from group.SPMContrasts import SPMContrasts
 from group.SPMResults import SPMResults
 
-from utility.images.Image import Image
-from utility.images.Images import Images
-from utility.images.transform_images import flirt
-from utility.images.images import mid_0based
-from utility.myfsl.utils.run import rrun
-from utility.matlab import call_matlab_spmbatch, call_matlab_function
-from utility.fileutilities import sed_inplace, copytree, get_filename
-from utility.utilities import is_list_of
+from myutility.images.Image import Image
+from myutility.images.Images import Images
+from myutility.images.transform_images import flirt
+from myutility.images.utilities import mid_0based
+from myutility.myfsl.utils.run import rrun
+from myutility.matlab import call_matlab_spmbatch, call_matlab_function
+from myutility.fileutilities import sed_inplace, copytree, get_filename
+from myutility.list import is_list_of
 
 
 class SubjectEpi:
